fixed routing. all protected routes begin with '/home'. switching to use redirect() instead of windows.location.href to avoid reloading whole page
