+++ src/app/api.ts
@@ -5,6 +5,7 @@ import {
   RegisterRequest,
 } from "../swagger/swagger";
 import { isProd } from "./utils/utils";
+import { redirect } from "next/navigation";
 
 const fetchWithBearer: typeof fetch = async (input, init = {}) => {
   const token = localStorage.getItem("token");
@@ -47,7 +48,7 @@ export const apiInstance = new Api({
 
 export const handleSignOut = () => {
   localStorage.clear();
-  window.location.href = "/";
+  redirect("/");
 };
 
 // API Formatting:
