+++ src/app/home/map/page.module.scss
@@ -1,4 +1,4 @@
-@use "../../styles/_variables.sass" as *;
+@use "../../../styles/_variables.sass" as *;
 
 .map_container {
   width: 100%;
