+++ src/app/home/profile/page.module.scss
@@ -1,0 +1,1 @@
+@use "../../../styles/_variables.sass" as *;
