--- src/app/profile/page.module.scss
@@ -1,1 +1,0 @@
-@use "../../styles/_variables.sass" as *;
